fuzzware-emulator/requirements.txt
setuptools
cython
PyYAML>=5.3
intelhex
archinfo==8.19.10.30
clint
capstone
IPython
ipdb
monkeyhex
pycapnp==v1.0.0b1